+++ arg_parser.py
@@ -10,6 +10,7 @@
         arg_cond_req
         arg_cond_req_or
         arg_default
+        arg_dir_chk
         arg_dir_chk_crt
         arg_file_chk
         arg_noreq_xor
@@ -174,6 +175,44 @@ def arg_default(arg, args_array, opt_def_dict):
 
     else:
         sys.exit("Error: Arg {0} missing value".format(arg))
+
+
+def arg_dir_chk(args_array, dir_perms_chk):
+
+    """Function:  arg_dir_chk
+
+    Description:  Checks to see if the directory has the correct permissions.
+
+    Arguments:
+        (input) args_array -> Array of command line options and values.
+        (input) dir_chk_perms -> Options with their directory perms in octal.
+        (output) status -> True|False - If directories have correct perms.
+
+    """
+
+    args_array = dict(args_array)
+    dir_perms_chk = dict(dir_perms_chk)
+    status = True
+
+    for item in set(dir_perms_chk) & set(args_array):
+
+        if not os.path.isdir(args_array[item]) or \
+           not os.access(args_array[item], os.X_OK):
+            print("Error: {0} does not exist or has permission denied.".format(
+                args_array[item]))
+            status = False
+
+        if gen_libs.octal_to_str(dir_perms_chk[item])[0] == "r" and \
+           not os.access(args_array[item], os.R_OK):
+            print("Error: {0} is not readable.".format(args_array[item]))
+            status = False
+
+        if gen_libs.octal_to_str(dir_perms_chk[item])[1] == "w" and \
+           not os.access(args_array[item], os.W_OK):
+            print("Error: {0} is not writable.".format(args_array[item]))
+            status = False
+
+    return status
 
 
 def arg_dir_chk_crt(args_array, dir_chk_list, dir_crt_list=None):