+++ test/unit/arg_parser/code_coverage.sh
@@ -13,6 +13,7 @@ coverage run -a --source=arg_parser test/unit/arg_parser/arg_add_def.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_cond_req.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_cond_req_or.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_default.py
+coverage run -a --source=arg_parser test/unit/arg_parser/arg_dir_chk.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_dir_chk_crt.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_file_chk.py
 coverage run -a --source=arg_parser test/unit/arg_parser/arg_noreq_xor.py